backend/pyproject.toml
[project]
name = "axon"
version = "0.0.0"  # Runtime version read from /app/VERSION
description = "Self-hosted AI command center — your personal chief of staff"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.0,<16",
    "litellm>=1.55.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "watchdog>=6.0.0",
    "tiktoken>=0.8.0",
    "httpx>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "cryptography>=43.0.0",
    "discord.py>=2.4.0",
    "slack-bolt>=1.20.0",
    "slack-sdk>=3.33.0",
    "pywebpush>=2.0.0",
    "structlog>=24.0.0",
    "lancedb>=0.6.0",
    "pyarrow>=15.0.0",
    "sentence-transformers>=3.0.0",
    "apscheduler>=3.10.0,<4",
]

[project.optional-dependencies]
postgres = [
    "asyncpg>=0.30.0",
]
voice = [
    "faster-whisper>=1.1.0",
    "silero-vad>=5.1",
    "piper-tts>=1.2.0",
]
web = [
    "trafilatura>=1.12.0",
]
media = [
    "youtube-transcript-api>=0.6.0",
]
browser = [
    "playwright>=1.40.0",
]
sandbox = [
    "docker>=7.0.0",
]
kubernetes = [
    "kubernetes>=28.0.0",
]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
